Solve ex 2.5.
I'm not thrilled about this it seems like there should be cleaner solution to getting a pair of trees that are almost the same size. On the other hand, I think this results in the maximum amount of sharing between the subtrees, so maybe this is as good as it gets.